requirements.txt
Django==5.1.1
djangorestframework==3.15.2
boto3==1.35.33
python-dotenv==1.0.1
pre-commit==4.0.0
